add port 444 for member virtual servers

+++ docs/files/docker-lauch.sh
@@ -1,5 +1,40 @@
 #!/bin/bash
 
+WIPE_ALL=0
+
+args=$(getopt -l "help" -l "wipe-all" -o "" -- "$@")
+
+eval set -- "$args"
+
+while [ $# -ge 1 ]; do
+    case "$1" in
+        --)
+            # No more options left.
+            shift
+            break
+            ;;
+        --wipe-all)
+            WIPE_ALL=1
+            ;;
+        -h|--help)
+            echo "$0: Launch the Byoda container"
+            echo ""
+            echo "Launches the Byoda container."
+            echo ""
+            echo "Usage: $0 [--help/-h] [--wipe-all]"
+            echo ""
+            echo "--help/-h     shows this helptext"
+            echo "--wipe-all    wipe all of the data of the pod and creates a new account ID before launching te container"
+            echo ""
+            return 0
+            ;;
+        *)
+           ;;
+    esac
+    shift
+done
+
+
 export BUCKET_PREFIX="changeme"             # Set to "IGNORE" when not using cloud storage
 export ACCOUNT_SECRET="changeme"            # Set to a long random string
 export PRIVATE_KEY_SECRET="changeme"        # set to long random string
@@ -9,6 +44,8 @@ if [[ "${BUCKET_PREFIX}" == "changeme" || "${ACCOUNT_SECRET}" == "changeme" || "
     exit 1
 fi
 
+ACCOUNT_FILE=~/.byoda-account_id
+
 DOCKER=$(which docker)
 
 if [ $? -ne 0 ]; then
@@ -16,7 +53,48 @@ if [ $? -ne 0 ]; then
     exit 1
 fi
 
-ACCOUNT_FILE=~/.byoda-account_id
+SYSTEM_MFCT=$(sudo dmidecode -t system | grep Manufacturer)
+SYSTEM_VERSION=$(sudo dmidecode -t system | grep Version)
+echo "System info:"
+echo "    ${SYSTEM_MFCT}"
+echo "    ${SYSTEM_VERSION}"
+
+if [[ "${SYSTEM_MFCT}" == *"Microsoft Corporation"* ]]; then
+    export CLOUD=Azure
+    echo "Running in cloud: ${CLOUD}"
+    if [[ "${WIPE_ALL}" == "1" ]]; then
+        echo "Wiping all data of the pod and creating a new account ID"
+        az storage blob delete-batch -s byoda --account-name ${BUCKET_PREFIX}private --auth-mode login
+        rm ${ACCOUNT_FILE}
+    fi
+elif [[ "${SYSTEM_MFCT}" == *"Google"* ]]; then
+    export CLOUD=GCP
+    echo "Running in cloud: ${CLOUD}"
+    if [[ "${WIPE_ALL}" == "1" ]]; then
+        echo "Wiping all data of the pod and creating a new account ID"
+        gcloud alpha storage rm --recursive gs://${BUCKET_PREFIX}-private/*
+        rm ${ACCOUNT_FILE}
+    fi
+elif [[ "${SYSTEM_VERSION}" == *"amazon"* ]]; then
+    export CLOUD=AWS
+    echo "Running in cloud: ${CLOUD}"
+    if [[ "${WIPE_ALL}" == "1" ]]; then
+        echo "Wiping all data of the pod and creating a new account ID"
+        aws s3 rm s3://${BUCKET_PREFIX}-private/private --recursive
+        aws s3 rm s3://${BUCKET_PREFIX}-private/network-byoda.net --recursive
+        rm ${ACCOUNT_FILE}
+    fi
+else
+    export CLOUD=LOCAL
+    echo "Not runing in a public cloud"
+    if [[ "${WIPE_ALL}" == "1" ]]; then
+        echo "Wiping all data of the pod and creating a new account ID"
+        sudo rm -rf ${ROOT_DIR} 2>/dev/null
+        sudo mkdir -p ${ROOT_DIR}
+        rm ${ACCOUNT_FILE}
+    fi
+fi
+
 if [ -f "${ACCOUNT_FILE}" ]; then
     ACCOUNT_ID=$(cat ${ACCOUNT_FILE})
     echo "Reading account_id from ${ACCOUNT_FILE}: ${ACCOUNT_ID}"
@@ -30,32 +108,6 @@ else
     echo "Writing account_id to ${ACCOUNT_FILE}: ${ACCOUNT_ID}"
 fi
 
-if [ -f ~/src/byoda-python/byoda/servers/pod_server.py ]; then
-    export PORT=$(grep HTTP_PORT ~/src/byoda-python/byoda/servers/pod_server.py | cut -d ' ' -f 7)
-else
-    export PORT=8000
-fi
-
-SYSTEM_MFCT=$(sudo dmidecode -t system | grep Manufacturer)
-SYSTEM_VERSION=$(sudo dmidecode -t system | grep Version)
-echo "System info:"
-echo "    ${SYSTEM_MFCT}"
-echo "    ${SYSTEM_VERSION}"
-
-if [[ "${SYSTEM_MFCT}" == *"Microsoft Corporation"* ]]; then
-    export CLOUD=Azure
-    echo "Running in cloud: ${CLOUD}"
-elif [[ "${SYSTEM_MFCT}" == *"Google"* ]]; then
-    export CLOUD=GCP
-    echo "Running in cloud: ${CLOUD}"
-elif [[ "${SYSTEM_VERSION}" == *"amazon"* ]]; then
-    export CLOUD=AWS
-    echo "Running in cloud: ${CLOUD}"
-else
-    export CLOUD=LOCAL
-    echo "Not runing in a public cloud"
-fi
-
 export ROOT_DIR=/byoda
 export LOGLEVEL=DEBUG
 export BOOTSTRAP=BOOTSTRAP
@@ -65,10 +117,12 @@ sudo mkdir -p ${LOGDIR}
 
 sudo docker stop byoda 2>/dev/null
 sudo docker rm byoda  2>/dev/null
-# sudo docker rmi byoda/byoda-pod:latest
 
-sudo rm -rf ${ROOT_DIR} 2>/dev/null
-sudo mkdir -p ${ROOT_DIR}
+if [[ "${CLOUD}" != "LOCAL" ]]; then
+    # Wipe the cache directory
+    sudo rm -rf ${ROOT_DIR} 2>/dev/null
+    sudo mkdir -p ${ROOT_DIR}
+fi
 
 echo "Creating container for account_id ${ACCOUNT_ID}"
 docker pull byoda/byoda-pod:latest
